Bug fixed: in Perl generated scripts, the "Perl die" tLogCatcher message of a tDie is no longer on the last component of the tLogCatcher flow, it is the current tDie, simply.
Bug fixed: in Perl generated scripts, the component execution duration was
always sent to tStatCatcher in case of failure, even if the "tStatCatcher
statistics" option had not been checked for the component. If the option had
not been checked, the duration was false. Now, the duration is calculated
only if the option is checked, as expected.


git-svn-id: http://talendforge.org/svn/tos/trunk@2881 f6f1c999-d317-4740-80b0-e6d1abc6f99e