feature 1681 added: in Perl tFileOutputPositional, user can choose the row separator. The related TUJ was added.
bug 1713 fixed: in Perl tFileOutputPositional, header titles are trunked to
the expected size of the data. The related TUJ was added.

Improvement: Perl tFileOutputPositional template code "was reduced to the
minimum" so that no useless code remain. encoding was removed because not
used yet and it may be confusing for users.


git-svn-id: http://talendforge.org/svn/tos/trunk@7294 f6f1c999-d317-4740-80b0-e6d1abc6f99e