fix(TESB-24920) (#3286)
* TESB-21832 Problems with job/route exec. with local runtime in Studio

Local runtime for Studio job/routes execution should not be used in some
cases: build type has to be checked to define the best execution
processor. As execution processor must change when build type is updated
in deployment tab, processor can not be persisted anymore.

* Studio graphical improvements

Fixed some graphical inconsistencies in Studio (borders, margins...)

* TESB-22230 NPE in build item wizard when changing build type

* TESB-22230 NPE in build item wizard when changing build type

* TESB-22230 NPE in build item wizard when changing build type

* TESB-24920 Default endpoint URI not added to endpoint with cRest

This commit improves the way endpoint URL is displayed in console when a
Route/Dataservice is ran in Studio (in case context variables are used).

* Revert "TESB-24920 Default endpoint URI not added to endpoint with cRest"

This reverts commit 0f43b7d4061a729a087ad1f2044d515f3130e6be.

* TESB-24920 Default endpoint URI not added to endpoint with cRest

This commit improves the way endpoint URL is displayed in console when a
Route/Dataservice is ran in Studio (in case context variables are used).